requirements.txt
python-dateutil
Pillow
psycopg2-binary
django-allauth
coverage